// compete-platform-front/app/_utils/types.tsx
export interface JoinToLobbyInfo {
  userId: number;
  inviterId?: number | null;
  teamId?: number | null;
  lobbyId: number;
  code?: string | null;
}
// ...
export interface LobbyAdminConfiguration {
  id: number;
  pickMaps: Map[];
  public: boolean;
  serverId: number;
  config: ServerConfig;
  playersAmount: Type;
  matchFormat: Format;
}

export interface GetUserBidDto {
  id: number;
  bid: number;
  userId: number;
}
export enum Map {
  Mirage,
  Inferno,
  Nuke,
  Anubis,
  Overpass,
  Vertigo,
  Ancient,
  Dust2,
  Office,
  Italy,
}
export interface GetServerDto {
  id: number;
  location: string;
  path: string;
  isHealthy: boolean;
}

export interface CancelLobbyRequest {
  id: number;
  offenders: number[] | null;
}
export interface ServerConfig {
  friendlyFire: boolean;
  freezeTime: number;
}

export enum Format {
  BO1 = 1,
  BO3 = 2,
  BO5 = 3,
}
// ...
export enum Type {
  v1 = 1,
  v2 = 2,
  v3 = 3,
  v4 = 4,
  v5 = 5,
}

export enum LobbyStatus {
  Configuring,
  Veto,
  Playing,
  Canceled,
  Over,
  Warmup,
}